pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "streamlit>=1.47.0",
    "uvicorn[standard]>=0.35.0",
    "pytesseract>=0.3.10",
]
